tests/submission/MOSQUITO/event-source/accept-header-include-text-event-stream.html

<!DOCTYPE html>
<html>
<head>
	<title>For HTTP connections, the Accept header may be included; if included, it must contain only formats of event framing that are supported by the user agent (one of which must be text/event-stream, as described below):</title>
  <meta charset="utf-8" />
  	<meta name=viewport content="width=device-width, maximum-scale=1.0, user-scalable=no" />
	  <link rel="author" title="Fabrice Clari" href="mailto:f.clari@inno-group.com">
	  <link rel="author" title="Dimitri Bocquet" href="mailto:Dimitri.Bocquet@mosquito-fp7.eu">
	<script src="/resources/testharness.js"></script>
	<script src="/resources/testharnessreport.js"></script>
	<script src="/resources/WebIDLParser.js"></script>
	<script src="/resources/idlharness.js"></script>
</head>
<body>
	<h1>For HTTP connections, the Accept header may be included; if included, it must contain only formats of event framing that are supported by the user agent (one of which must be text/event-stream, as described below):</h1>
	<div id="log"></div>
  <script>
    var s = new EventSource('http://html5.mosquito-fp7.eu/messages/t0030');
    ttext = async_test("accept header include 'text/event-stream'");

    s.onopen = function(e){ 
    	/* Connection opened. */
    };
    	
    s.onmessage = function(e){
    console.log(e.data);
    	var dat = e.data.split("\n");
    	ttext.step(function(){assert_equals(dat[1].toString(), "text/event-stream")});
    	ttext.done();
    };
    	
    s.onerror = function(e){
    	/* Connection closed. */
    };

  </script>
  </div>
</body>
</html>
